Rename base dir
